+++ filename.js
@@ -1,3 +1,3 @@
-const FileNamePlugin = require('./build/file_name_plugin');
+const FileNamePlugin = require('./build/file_name_plugin/plugin');
 
 module.exports = FileNamePlugin;
