+++ package.json
@@ -10,18 +10,15 @@
     "url": "https://github.com/jest-community/jest-watch-typeahead.git"
   },
   "homepage": "https://github.com/jest-community/jest-watch-typeahead",
-  "files": [
-    "build/",
-    "filename.js",
-    "testname.js"
-  ],
+  "files": ["build/", "filename.js", "testname.js"],
   "scripts": {
     "test": "jest",
     "lint": "eslint .",
     "watch": "babel src -w --ignore **/*.test.js,integration -d build",
     "build": "babel src --ignore **/*.test.js,integration -d build",
     "prepublish": "yarn build",
-    "format": "prettier --single-quote --trailing-comma all --write \"!(build)/**/*.js\""
+    "format":
+      "prettier --single-quote --trailing-comma all --write \"!(build)/**/*.js\""
   },
   "dependencies": {
     "chalk": "^2.4.1",
@@ -47,6 +44,7 @@
     "prettier": "1.11.1"
   },
   "jest": {
+    "watchPlugins": ["<rootDir>/filename", "<rootDir>/testname"],
     "snapshotSerializers": [
       "<rootDir>/node_modules/pretty-format/build/plugins/convert_ansi"
     ],
