+++ src/lib/pattern_mode_helpers.js
