+++ testname.js
@@ -1,3 +1,3 @@
-const TestNamePlugin = require('./build/test_name_plugin');
+const TestNamePlugin = require('./build/test_name_plugin/plugin');
 
 module.exports = TestNamePlugin;
